feat: order documents cleanup
