--- apps/web/components/DocumentsList/types.ts
@@ -1,6 +1,0 @@
-import { OrderDocument } from '@plentymarkets/shop-api';
-
-export type DocumentsListProps = {
-  documents: OrderDocument[];
-  accessKey: string;
-};
