+++ apps/web/components/OrderDocumentsList/types.ts
@@ -1,0 +1,5 @@
+import { Order } from '@plentymarkets/shop-api';
+
+export type DocumentsListProps = {
+  order: Order;
+};
